Refactored (moved).
