+++ demeter-base/src/test/java/com/heliosphere/demeter/base/DemeterTestSuite.java
@@ -17,8 +17,8 @@ import org.junit.runners.Suite.SuiteClasses;
 
 import com.heliosphere.demeter.base.file.xml.XmlFileTest;
 import com.heliosphere.demeter.base.file.xml.XmlRunnerParameterFileTest;
-import com.heliosphere.demeter.base.resource.bundle.test.ResourceBundleTest;
-import com.heliosphere.demeter.base.resource.test.ResourceTest;
+import com.heliosphere.demeter.base.resource.ResourceTest;
+import com.heliosphere.demeter.base.resource.bundle.ResourceBundleTest;
 
 /**
  * Defines the {@code demeter} unit test suite.
