+++ demeter-base/src/test/java/com/heliosphere/demeter/base/resource/ResourceTest.java
@@ -9,7 +9,7 @@
  * License can be consulted at http://www.apache.org/licenses/LICENSE-2.0
  * ---------------------------------------------------------------------------
  */
-package com.heliosphere.demeter.base.resource.test;
+package com.heliosphere.demeter.base.resource;
 
 import static org.junit.Assert.fail;
